configstore: split store.go into cohesive same-package files (#2158)
store.go had grown to 2112 LOC, over the ~2000 production-LOC modularity
threshold (docs/engineering-style.md). Split it into six cohesive files
in the same package (configstore), pure code-motion with no logic change,
no exported-API change, and byte-identical runtime behavior. This is the
P1 Go file-split from the converged #2158 plan, section 5.3.

The boundary follows the plan recommendation:

  store.go         Store struct, New, SetConfigDBWriterVersion, the
                   node/cluster accessors (SetNodeID, Set/ClusterReadOnly),
                   the compile + schema-validate pipeline (compileTree*,
                   schemaValidateExpandedTree*), and SyncApply.
  store_persist.go Load/Save, writeActive/writeActiveMarker, journalLog,
                   journalConfigHash, the #1799 degrade-and-retry persist
                   machinery (ConfigPersistDegraded,
                   noteActivePersistFailureLocked, persistRetryLoop),
                   config archival, and rescue config.
  store_lock.go    config-mode enter/exit locking (EnterConfigure*,
                   ExitConfigure*, ForceExitConfigure, ConfigHolder,
                   IsExclusiveLocked) and edit-path navigation.
  store_command.go candidate edit verbs (Set, Delete, Copy, Rename,
                   Insert, Annotate, Deactivate/Activate, Load*) and the
                   centralized flat-line replay (applyEditLine, LoadSet).
  store_commit.go  commit / commit-confirmed / rollback, the confirm-timer
                   machinery (SetRollbackExecutor, fireConfirmTimer,
                   PromoteRollback, performAutoRollback), commit history
                   (ListHistory, ListCommitHistory, CommitDiffSummary), and
                   rollback-history file persistence (rollbackPath,
                   saveRollbackFiles, cleanupRollbackFiles,
                   loadRollbackHistory).
  store_format.go  the Show* render family (Candidate/Active/Rollback x
                   Set/JSON/XML/Inheritance), ShowCompare*, ExportJSON, and
                   the read-only accessors (ActiveConfig, ActiveTree,
                   EverCommitted, CompileCandidate, splitLines).

Resulting per-file LOC: store.go 419, store_persist.go 378, store_lock.go
165, store_command.go 324, store_commit.go 580, store_format.go 307 — all
well under the threshold.

Methodology checks (the code-motion bar, plan section 8):
  - Move-only: the set of top-level func/method signatures is byte
    identical before and after (103 funcs, identical sorted set); a
    line-level diff of all non-import/non-package body lines shows zero
    lines removed and only the new package-doc comment added.
  - No API diff: `go doc -all ./pkg/configstore/` is byte identical
    before and after.
  - No init-order risk: store.go has no package-level var initializer or
    init() function, so source-file-name ordering is irrelevant.
  - Receivers preserved: every moved method keeps its exact (s *Store)
    receiver; no function silently became free-standing.
  - Build + test green: go build ./..., go vet ./pkg/configstore/, and
    go test ./pkg/configstore/... ./pkg/daemon/... all pass unchanged
    from master.
