+++ pkg/configstore/store.go
@@ -1,23 +1,32 @@
 // Package configstore implements the Junos-style candidate/active
 // configuration management with commit and rollback support.
+//
+// The Store type is split across several same-package files for
+// readability (#2158 code-motion, no behavior change):
+//   - store.go         — Store struct, New, node/cluster accessors, the
+//                         compile/schema-validate pipeline, SyncApply
+//   - store_persist.go — Load/Save, writeActive*, journal helpers, the
+//                         #1799 degrade-and-retry persist machinery,
+//                         config archival, and rescue config
+//   - store_lock.go    — config-mode enter/exit locking + edit-path nav
+//   - store_command.go — candidate edit verbs (set/delete/copy/...) and
+//                         the flat-line replay (LoadSet/LoadMerge/...)
+//   - store_commit.go  — commit / commit-confirmed / rollback, the
+//                         confirm-timer machinery, and rollback-history
+//                         file persistence
+//   - store_format.go  — the Show* render family + read-only accessors
 package configstore
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"log/slog"
-	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/psaab/xpf/pkg/config"
 	"github.com/psaab/xpf/pkg/configstore/journal"
-	"github.com/psaab/xpf/pkg/fsatomic"
 )
 
 // JournalEntry is the audit-log record type, owned by the journal
@@ -167,255 +176,6 @@ func New(filePath string) (*Store, error) {
 // path is the only caller.
 func (s *Store) SetConfigDBWriterVersion(v string) {
 	s.db.SetWriterVersion(v)
-}
-
-// Load builds the configuration from disk.
-func (s *Store) Load() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	tree, committed, err := s.db.ReadActiveMeta()
-	if err != nil {
-		// A read/parse/decrypt/envelope failure on a PRESENT active.json is
-		// the fail-closed case (#1917 increment B, D1): an unparseable or
-		// too-new DB must NOT be silently overwritten by a blind bootstrap.
-		// Tag it with ErrConfigDBUnreadable so the daemon can make it fatal
-		// (daemon_run.go), distinct from a compile error (handled leniently
-		// below) or an absent DB (handled above as start-fresh).
-		return fmt.Errorf("read config: %w: %w", ErrConfigDBUnreadable, err)
-	}
-	if tree == nil {
-		// Absent DB: start fresh. everCommitted stays false (a never-booted
-		// store has never committed); the daemon's bootstrapFromFile may
-		// import a preseeded xpf.conf, which resolves NOT-bootstrap on its
-		// own (case 2). The #1922 step-0 marker only governs the DB-present
-		// disambiguation.
-		return nil
-	}
-	// #1922 step-0 marker: record whether the on-disk DB represents a
-	// successfully-committed config. A legacy/older-build DB (no envelope
-	// field) reads committed=true (migration rule C3), so an upgrade never
-	// misclassifies an existing active config into bootstrap.
-	s.everCommitted = committed
-	// Seed the degraded-retry marker from the on-disk state too (Copilot
-	// finding): if Load reads a never-committed DB (committed=0) and a later
-	// persist failure triggers the #1799 retry loop BEFORE any commit/sync
-	// resets the marker, the retry must re-write committed=0 — not the
-	// New() default of true, which would silently heal the never-committed
-	// marker into an operator-committed-empty DB and re-enable takeover.
-	s.persistMarkerCommitted = committed
-
-	// Rolling-upgrade tolerance (#1373 / #1525): a node may boot
-	// with `system dataplane-type ebpf` or `... dpdk` persisted
-	// from before the retirement-strict validator landed. Without
-	// this rewrite, compileTree below returns
-	// ErrEBPFDataplaneRetired / ErrDPDKDataplaneRetired, the daemon
-	// gets nil active config, and bootstraps blind. Rewriting the
-	// leaf to absent (defaults to userspace) lets the daemon come
-	// up so the operator can fix the config from CLI.
-	rewriteRetiredDataplaneType(tree, LoadCaller)
-
-	// #1798 migration: a persisted free-text value carrying control
-	// characters (e.g. a "lan\nDHCP=ipv4" description committed before
-	// the strict commit-time gate landed) must neither fail boot now
-	// nor make the operator's next unrelated commit fail mysteriously.
-	// Scrub the tree in place with a warning — this tree becomes the
-	// active config (and the candidate clones from it), so the next
-	// strict commit sees only clean values.
-	for _, p := range config.SanitizeTreeControlChars(tree) {
-		slog.Warn("sanitized control characters in persisted config value",
-			"path", p, "issue", "#1798")
-	}
-
-	// Tolerant compile: an already-persisted config must boot through
-	// (see compileTreeLenient for the validator downgrades).
-	compiled, err := s.compileTreeLenient(tree)
-	if err != nil {
-		// #1960 fail-closed: the bytes read+parsed fine (this is NOT
-		// ErrConfigDBUnreadable) but a PRESENT, previously-committed config
-		// no longer compiles. s.everCommitted was already set true above, so
-		// the daemon could otherwise see ActiveConfig()==nil + everCommitted
-		// and resolve to NORMAL boot — positional claim-all interface naming.
-		// Tag the error with ErrConfigCompile so the daemon can detect this
-		// edge with errors.Is and refuse takeover (enter the #1922
-		// bootstrap/lifeline safe state) instead.
-		//
-		// s.compiled MUST stay nil — ActiveConfig() returns s.compiled, and
-		// nil is precisely the signal that forces bootstrap (computeBootClass).
-		// BUT retain the parsed-but-broken tree as s.active and load rollback
-		// history, so the recovery the daemon advertises ("fix the config from
-		// the CLI/gRPC and commit, or roll back") actually works (Codex #1991
-		// r1): EnterConfigure clones s.active into the candidate, so `configure`
-		// + `show | compare` surface the broken stanza for the operator to fix,
-		// and Rollback(n) / `show | compare rollback n` reach the on-disk
-		// history. Without this, s.active stayed the empty New() tree and the
-		// history was never loaded — the operator saw an empty config and no
-		// rollbacks, with no in-band way to recover. s.active is always non-nil
-		// (New seeds an empty tree), so the (active non-nil, compiled nil) shape
-		// here is the same one a fresh boot already has — no new invariant.
-		s.active = tree
-		s.loadRollbackHistory()
-		return fmt.Errorf("compile config: %w: %w", ErrConfigCompile, err)
-	}
-
-	s.active = tree
-	s.compiled = compiled
-	s.loadRollbackHistory()
-	return nil
-}
-
-// Save persists the active configuration to disk.
-func (s *Store) Save() error {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	return s.writeActive(s.active)
-}
-
-// writeActive persists tree as the on-disk active configuration.
-// Routes through the writeActiveFn test seam when set (#1799);
-// otherwise uses the DB's durable temp + fsync + rename + dir-fsync
-// write (#1894). Caller must hold s.mu (read or write lock).
-func (s *Store) writeActive(tree *config.ConfigTree) error {
-	if s.writeActiveFn != nil {
-		return s.writeActiveFn(tree)
-	}
-	return s.db.WriteActive(tree)
-}
-
-// writeActiveMarker persists tree as the on-disk active config with an
-// explicit #1922 step-0 committed marker. committed=false writes the
-// never-committed marker (Item 1b first-commit rollback). Routes through
-// the writeActiveMarkerFn test seam when set; otherwise the production
-// DB.WriteActiveMarker. Caller must hold s.mu. When only the legacy
-// writeActiveFn seam is set (older tests), the marker degrades to that seam
-// (the committed bit is not observable through it, which is acceptable for
-// those tests — the marker behavior has dedicated coverage via the DB seam).
-func (s *Store) writeActiveMarker(tree *config.ConfigTree, committed bool) error {
-	if s.writeActiveMarkerFn != nil {
-		return s.writeActiveMarkerFn(tree, committed)
-	}
-	if s.writeActiveFn != nil {
-		return s.writeActiveFn(tree)
-	}
-	return s.db.WriteActiveMarker(tree, committed)
-}
-
-// journalLog appends an audit entry; a failure is surfaced as a
-// warning only — journaling must never fail a commit/sync/rollback,
-// and the persist_error path must not recurse into journaling its own
-// failure (#1896).
-func (s *Store) journalLog(e *JournalEntry) {
-	if err := s.journal.Log(e); err != nil {
-		slog.Warn("config journal append failed", "action", e.Action, "err", err)
-	}
-}
-
-// journalConfigHash returns the sha256 hex of the tree's Format() text
-// — the same text saveRollbackFiles writes to rollback slots, so a
-// retained rollback file can be correlated to its journal entry with
-// `sha256sum` (#1896). Best-effort correlation: slots shift on every
-// commit and only history.MaxSize() of them are kept.
-func journalConfigHash(tree *config.ConfigTree) string {
-	if tree == nil {
-		return ""
-	}
-	sum := sha256.Sum256([]byte(tree.Format()))
-	return hex.EncodeToString(sum[:])
-}
-
-// ConfigPersistDegraded reports whether the running active config
-// failed to persist to disk on an Option-B path (SyncApply /
-// performAutoRollback) and the background retry has not yet succeeded
-// (#1799). While true, a daemon restart would load a STALE config;
-// /health returns 503 and xpf_daemon_config_persist_degraded reads 1.
-func (s *Store) ConfigPersistDegraded() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.persistDegraded
-}
-
-// noteActivePersistFailureLocked records a WriteActive failure on an
-// Option-B path (#1799): the in-memory apply already happened (HA
-// convergence / auto-rollback safety win over disk trouble), so the
-// failure must become visible and self-healing instead of fatal.
-// Flips the degraded flag, writes a journal ERROR entry, and starts
-// the singleton retry goroutine. Must be called with s.mu held
-// (write lock).
-func (s *Store) noteActivePersistFailureLocked(action string, err error) {
-	slog.Error("active config persist failed — running config is not durable; restart would load stale config",
-		"action", action, "err", err, "issue", "#1799")
-	s.persistDegraded = true
-	// The retry loop re-writes s.active with the marker last requested by the
-	// failing path. Callers that need the never-committed marker (Item 1b)
-	// set s.persistMarkerCommitted=false BEFORE invoking this; all other
-	// paths leave it at the default true.
-	s.journalLog(&JournalEntry{
-		Action: "persist_error",
-		Detail: fmt.Sprintf("%s: write active config failed: %v", action, err),
-	})
-	if s.persistRetryActive {
-		return
-	}
-	s.persistRetryActive = true
-	initial := s.persistRetryInitialBackoff
-	if initial <= 0 {
-		initial = time.Second
-	}
-	maxBackoff := s.persistRetryMaxBackoff
-	if maxBackoff <= 0 {
-		maxBackoff = 60 * time.Second
-	}
-	go s.persistRetryLoop(initial, maxBackoff)
-}
-
-// persistRetryLoop retries persisting the CURRENT active config with
-// doubling backoff until a write succeeds or the degraded flag is
-// cleared by a successful write on another path (#1799). Each attempt
-// re-reads s.active under s.mu and writes while holding the lock —
-// never a stale captured tree — so rapid successive syncs cannot
-// persist out-of-order state (the commit paths themselves write under
-// s.mu, so lock-held writes are the existing serialization).
-//
-// Shutdown safety: the Store has no close signal, and none is needed.
-// This is a plain goroutine outside any WaitGroup — it sleeps between
-// attempts and holds s.mu only for the duration of one atomic
-// temp-file write, so it cannot block daemon shutdown; process exit
-// simply abandons it.
-func (s *Store) persistRetryLoop(backoff, maxBackoff time.Duration) {
-	for {
-		time.Sleep(backoff)
-		s.mu.Lock()
-		if !s.persistDegraded {
-			// A successful write on a commit/sync path already
-			// persisted the current active config.
-			s.persistRetryActive = false
-			s.mu.Unlock()
-			return
-		}
-		// #1922: re-write with the marker the failing path requested
-		// (committed=false only for a first-commit rollback). For every
-		// other path persistMarkerCommitted is true, so this matches the
-		// pre-#1922 committed=1 write exactly.
-		err := s.writeActiveMarker(s.active, s.persistMarkerCommitted)
-		if err == nil {
-			s.persistDegraded = false
-			s.persistRetryActive = false
-			s.journalLog(&JournalEntry{
-				Action: "persist_recovered",
-				Detail: "active config persisted after earlier write failure",
-			})
-			s.mu.Unlock()
-			slog.Info("active config persisted after earlier write failure", "issue", "#1799")
-			return
-		}
-		s.mu.Unlock()
-		slog.Warn("active config persist retry failed", "err", err, "retry_in", backoff*2)
-		backoff *= 2
-		if backoff > maxBackoff {
-			backoff = maxBackoff
-		}
-	}
 }
 
 // SetClusterReadOnly toggles cluster read-only mode. When enabled, config
@@ -656,1457 +416,4 @@ func (s *Store) SyncApply(content string, chassisPreserve func(*config.ConfigTre
 
 	s.saveRollbackFiles()
 	return compiled, nil
-}
-
-// EnterConfigure enters configuration mode by cloning the active config.
-// Returns an error if another session is already in config mode.
-func (s *Store) EnterConfigure() error {
-	return s.EnterConfigureSession("")
-}
-
-// EnterConfigureSession enters configuration mode with a session identifier.
-// If the same session already holds the lock, it's a no-op.
-func (s *Store) EnterConfigureSession(sessionID string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.clusterReadOnly {
-		return fmt.Errorf("configuration database is not writable (secondary node)")
-	}
-	if s.configDir {
-		// Allow re-entry by same session.
-		if sessionID != "" && s.configHolder == sessionID {
-			return nil
-		}
-		// #2157: return the ErrConfigLocked sentinel (wrapping the plain
-		// message) so deferrable callers (the event-options action worker)
-		// can errors.Is it and retry instead of string-matching.
-		return fmt.Errorf("%w", ErrConfigLocked)
-	}
-	s.candidate = s.active.Clone()
-	s.configDir = true
-	s.dirty = false
-	s.configHolder = sessionID
-	s.configLockAt = time.Now()
-	return nil
-}
-
-// EnterConfigureExclusive enters exclusive configuration mode.
-func (s *Store) EnterConfigureExclusive(holder string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.clusterReadOnly {
-		return fmt.Errorf("configuration database is not writable (secondary node)")
-	}
-	if s.configDir {
-		return fmt.Errorf("%w", ErrConfigLocked)
-	}
-	s.candidate = s.active.Clone()
-	s.configDir = true
-	s.dirty = false
-	s.exclusiveHolder = holder
-	s.configLockAt = time.Now()
-	return nil
-}
-
-// ExitConfigureSession exits configuration mode only if the given session holds
-// the lock. Returns true if the lock was released.
-func (s *Store) ExitConfigureSession(sessionID string) bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if !s.configDir {
-		return false
-	}
-	if sessionID != "" && s.configHolder != sessionID {
-		return false
-	}
-	s.candidate = nil
-	s.configDir = false
-	s.dirty = false
-	s.exclusiveHolder = ""
-	s.configHolder = ""
-	s.editPath = nil
-	return true
-}
-
-// ForceExitConfigure exits configuration mode regardless of who holds the lock.
-// Used for stale lock cleanup.
-func (s *Store) ForceExitConfigure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if !s.configDir {
-		return
-	}
-	slog.Warn("force-releasing stale config lock", "holder", s.configHolder,
-		"held_for", time.Since(s.configLockAt).Round(time.Second))
-	s.candidate = nil
-	s.configDir = false
-	s.dirty = false
-	s.exclusiveHolder = ""
-	s.configHolder = ""
-	s.editPath = nil
-}
-
-// ConfigHolder returns the session ID of the current config lock holder
-// and whether the lock is held.
-func (s *Store) ConfigHolder() (string, bool) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.configHolder, s.configDir
-}
-
-// IsExclusiveLocked returns true if exclusive mode is active.
-func (s *Store) IsExclusiveLocked() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.exclusiveHolder != ""
-}
-
-// ExitConfigure exits configuration mode, discarding the candidate.
-func (s *Store) ExitConfigure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.candidate = nil
-	s.configDir = false
-	s.dirty = false
-	s.exclusiveHolder = ""
-	s.editPath = nil
-}
-
-// SetEditPath sets the edit path for hierarchical navigation.
-func (s *Store) SetEditPath(path []string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.editPath = path
-}
-
-// GetEditPath returns a copy of the current edit path.
-func (s *Store) GetEditPath() []string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return append([]string{}, s.editPath...)
-}
-
-// NavigateUp moves the edit path up one level.
-func (s *Store) NavigateUp() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if len(s.editPath) > 0 {
-		s.editPath = s.editPath[:len(s.editPath)-1]
-	}
-}
-
-// NavigateTop resets the edit path to the root.
-func (s *Store) NavigateTop() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.editPath = nil
-}
-
-// InConfigMode returns true if currently in configuration mode.
-func (s *Store) InConfigMode() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.configDir
-}
-
-// IsDirty returns true if the candidate differs from active.
-func (s *Store) IsDirty() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.dirty
-}
-
-// Set applies a "set" command to the candidate configuration.
-func (s *Store) Set(path []string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	if err := s.candidate.SetPath(path); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// SetFromInput parses a "set ..." command string and applies it.
-func (s *Store) SetFromInput(input string) error {
-	path, err := config.ParseSetCommand("set " + input)
-	if err != nil {
-		return err
-	}
-	return s.Set(path)
-}
-
-// Delete removes a node at the given path from the candidate configuration.
-func (s *Store) Delete(path []string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	if err := s.candidate.DeletePath(path); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// DeleteFromInput parses a "delete ..." command string and applies it.
-func (s *Store) DeleteFromInput(input string) error {
-	path, err := config.ParseSetCommand("delete " + input)
-	if err != nil {
-		return err
-	}
-	return s.Delete(path)
-}
-
-// DeactivateFromInput marks the candidate node at the given path inactive
-// (#2051), implementing the interactive Junos `deactivate <path>` verb. input
-// is the bare path WITHOUT the verb (mirroring SetFromInput/DeleteFromInput).
-//
-// It routes through applyEditLine — the same centralized verb switch used by
-// LoadSet / LoadMerge flat-line replay (store.go applyEditLine) — so the verb
-// logic lives in exactly one place. It deliberately does NOT go through
-// ParseSetCommand("set "+input): that parser would build the junk path
-// "deactivate <path>" (a config node literally named "deactivate"), never
-// reaching DeactivatePath. The node must already exist; DeactivatePath on an
-// already-inactive node is idempotent (it re-sets a bool).
-func (s *Store) DeactivateFromInput(input string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-	if err := applyEditLine(s.candidate, "deactivate "+input); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// ActivateFromInput clears the inactive marker on the candidate node at the
-// given path (#2051), implementing the interactive Junos `activate <path>`
-// verb. Symmetric with DeactivateFromInput; ActivatePath on an already-active
-// node is idempotent.
-func (s *Store) ActivateFromInput(input string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-	if err := applyEditLine(s.candidate, "activate "+input); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// Copy duplicates a config subtree from srcPath to dstPath.
-func (s *Store) Copy(srcPath, dstPath []string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-	if err := s.candidate.CopyPath(srcPath, dstPath); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// Rename moves a config subtree from srcPath to dstPath.
-func (s *Store) Rename(srcPath, dstPath []string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-	if err := s.candidate.RenamePath(srcPath, dstPath); err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// Insert moves an element before or after a reference element within the
-// same parent's ordered children list.
-func (s *Store) Insert(elementPath, refPath []string, before bool) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-	var err error
-	if before {
-		err = s.candidate.InsertBefore(elementPath, refPath)
-	} else {
-		err = s.candidate.InsertAfter(elementPath, refPath)
-	}
-	if err != nil {
-		return err
-	}
-	s.dirty = true
-	return nil
-}
-
-// Annotate sets a comment on a configuration node in the candidate config.
-func (s *Store) Annotate(path []string, comment string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	children := s.candidate.Children
-	var target *config.Node
-	for _, key := range path {
-		found := false
-		for _, child := range children {
-			for _, k := range child.Keys {
-				if k == key {
-					target = child
-					children = child.Children
-					found = true
-					break
-				}
-			}
-			if found {
-				break
-			}
-		}
-		if !found {
-			return fmt.Errorf("path not found: %s", strings.Join(path, " "))
-		}
-	}
-
-	target.Annotation = comment
-	s.dirty = true
-	return nil
-}
-
-// LoadOverride replaces the entire candidate config with the parsed input.
-// The input can be hierarchical Junos config or flat "set" commands.
-func (s *Store) LoadOverride(content string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	tree, errs := config.NewParser(content).Parse()
-	if len(errs) > 0 {
-		return fmt.Errorf("parse error: %v", errs[0])
-	}
-
-	s.candidate = tree
-	s.dirty = true
-	return nil
-}
-
-// LoadMerge merges the parsed input into the existing candidate config.
-// For flat "set" commands, each line is applied individually.
-// For hierarchical input, it's converted to set commands and merged.
-func (s *Store) LoadMerge(content string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	// Detect format: if content has set/delete/deactivate/activate lines,
-	// process as flat command lines.
-	lines := strings.Split(content, "\n")
-	isSetFormat := false
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		if strings.HasPrefix(trimmed, "set ") ||
-			strings.HasPrefix(trimmed, "delete ") ||
-			strings.HasPrefix(trimmed, "deactivate ") ||
-			strings.HasPrefix(trimmed, "activate ") {
-			isSetFormat = true
-			break
-		}
-	}
-
-	if isSetFormat {
-		for _, line := range lines {
-			trimmed := strings.TrimSpace(line)
-			if trimmed == "" || strings.HasPrefix(trimmed, "#") {
-				continue
-			}
-			if err := applyEditLine(s.candidate, trimmed); err != nil {
-				return fmt.Errorf("line %q: %w", trimmed, err)
-			}
-		}
-	} else {
-		// Parse as hierarchical config and merge each top-level node
-		tree, errs := config.NewParser(content).Parse()
-		if len(errs) > 0 {
-			return fmt.Errorf("parse error: %v", errs[0])
-		}
-		// Convert hierarchical to flat commands and apply each one. FormatSet
-		// emits a `deactivate <path>` line after every inactive node's `set`
-		// line(s), so the hierarchical -> flat -> tree round trip must honor
-		// the deactivate verb (#2008 H1) to preserve Inactive — applying it as
-		// a plain set would silently re-activate the node.
-		setLines := strings.Split(tree.FormatSet(), "\n")
-		for _, line := range setLines {
-			trimmed := strings.TrimSpace(line)
-			if trimmed == "" {
-				continue
-			}
-			if err := applyEditLine(s.candidate, trimmed); err != nil {
-				return fmt.Errorf("merge: %w", err)
-			}
-		}
-	}
-
-	s.dirty = true
-	return nil
-}
-
-// applyEditLine parses a single flat command line and applies the correct
-// edit to tree based on its verb (#2008 H1): set, delete, deactivate, or
-// activate. Centralizing the verb switch keeps every flat-line replay path
-// (LoadMerge flat + hierarchical-via-FormatSet, LoadSet) in agreement, so
-// `show | display set` output — which emits `deactivate <path>` for inactive
-// nodes — round-trips back to an inactive node instead of being skipped (and
-// reloaded active) or parsed as a junk path literally starting "deactivate".
-func applyEditLine(tree *config.ConfigTree, line string) error {
-	verb, path, err := config.ParseSetVerb(line)
-	if err != nil {
-		return err
-	}
-	switch verb {
-	case "delete":
-		return tree.DeletePath(path)
-	case "deactivate":
-		return tree.DeactivatePath(path)
-	case "activate":
-		return tree.ActivatePath(path)
-	default: // "set" (or a bare, unprefixed path)
-		return tree.SetPath(path)
-	}
-}
-
-// LoadSet applies multiple flat command lines to the candidate config.
-// Each line starting with a recognized verb — set, delete, deactivate, or
-// activate (#2008 H1) — is parsed and applied; other lines (e.g. comments)
-// are skipped. The deactivate/activate verbs make `show | display set`
-// output round-trippable: previously a `deactivate <path>` line was skipped
-// here, so an inactive node reloaded ACTIVE.
-func (s *Store) LoadSet(content string) (int, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.candidate == nil {
-		return 0, fmt.Errorf("not in configuration mode")
-	}
-	count := 0
-	for _, line := range strings.Split(content, "\n") {
-		line = strings.TrimSpace(line)
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		if !strings.HasPrefix(line, "set ") &&
-			!strings.HasPrefix(line, "delete ") &&
-			!strings.HasPrefix(line, "deactivate ") &&
-			!strings.HasPrefix(line, "activate ") {
-			continue
-		}
-		if err := applyEditLine(s.candidate, line); err != nil {
-			return count, fmt.Errorf("line %q: %w", line, err)
-		}
-		count++
-	}
-	s.dirty = true
-	return count, nil
-}
-
-// CommitCheck validates the candidate configuration without applying it.
-func (s *Store) CommitCheck() (*config.Config, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	if s.candidate == nil {
-		return nil, fmt.Errorf("not in configuration mode")
-	}
-
-	compiled, err := s.compileTree(s.candidate)
-	if err != nil {
-		return nil, err
-	}
-
-	return compiled, nil
-}
-
-// Commit validates, compiles, and applies the candidate configuration.
-// Returns the compiled config for the caller to apply to the dataplane.
-// Identical to CommitWithDescription with an empty description (the two
-// were verbatim duplicates before #1799 unified them).
-func (s *Store) Commit() (*config.Config, error) {
-	return s.CommitWithDescription("")
-}
-
-// CommitWithDescription validates, compiles, and applies the candidate configuration
-// with an optional comment/description attached to the history and journal entries.
-//
-// Persistence contract (#1799, Option A — persist-before-promote): the
-// candidate tree is written to the on-disk active config BEFORE any
-// in-memory promotion. WriteActive is temp-file + rename atomic
-// (db.go), so a persist failure leaves the previous active config
-// intact on disk; the commit then fails with the candidate left
-// intact and NOTHING mutated — no active/candidate/compiled/dirty
-// change, no history push, no journal entry, no rollback-file save.
-// A commit that reports success can therefore never silently revert
-// to the previous config on daemon restart.
-func (s *Store) CommitWithDescription(description string) (*config.Config, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return nil, fmt.Errorf("not in configuration mode")
-	}
-
-	compiled, err := s.compileTree(s.candidate)
-	if err != nil {
-		return nil, fmt.Errorf("commit check failed: %w", err)
-	}
-
-	// #1799 Option A: persist BEFORE promote. On failure the old
-	// active stays on disk and in memory; the operator sees the
-	// error and the candidate is still there to retry.
-	if err := s.writeActive(s.candidate); err != nil {
-		return nil, fmt.Errorf("commit failed: persist active config: %w", err)
-	}
-	s.persistDegraded = false       // disk now holds the current config
-	s.everCommitted = true          // #1922 step-0: a real commit has succeeded
-	s.persistMarkerCommitted = true // #1922: degraded-retry writes committed=1
-
-	// Push current active to history with description
-	s.history.Push(&HistoryEntry{
-		Config:    s.active.Clone(),
-		Timestamp: time.Now(),
-		Comment:   description,
-	})
-
-	// Promote candidate to active
-	s.active = s.candidate
-	s.candidate = s.active.Clone()
-	s.compiled = compiled
-	s.dirty = false
-
-	// Log to journal with description
-	s.journalLog(&JournalEntry{
-		Action:     "commit",
-		Detail:     description,
-		ConfigHash: journalConfigHash(s.active),
-	})
-
-	s.saveRollbackFiles()
-
-	// Auto-archive if configured
-	if s.archiveDir != "" {
-		max := s.archiveMax
-		if max <= 0 {
-			max = 10
-		}
-		go func() {
-			if err := s.ArchiveConfig(s.archiveDir, max); err != nil {
-				slog.Warn("auto-archive failed", "err", err)
-			}
-		}()
-	}
-
-	return compiled, nil
-}
-
-// SetRollbackExecutor registers the daemon's commit-confirmed timeout
-// rollback transaction (#1922 Item 1a). The executor is invoked (on the
-// timer's own goroutine, never under s.mu) with the confirm generation
-// that armed the timer. It MUST acquire the daemon apply semaphore
-// first, then call PromoteRollback(gen) + re-apply the returned config,
-// so store promotion and dataplane re-apply are atomic under the same
-// lock that serializes commits. Registering nil disables the executor
-// and reverts to the self-contained performAutoRollback fallback.
-func (s *Store) SetRollbackExecutor(fn func(gen uint64)) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.rollbackExecutor = fn
-}
-
-// CommitConfirmed validates, compiles, and applies the candidate with an
-// automatic rollback timer. If minutes is 0, defaults to 10.
-// If a bare "commit" is not issued within the timeout, the config auto-reverts.
-//
-// Persistence contract (#1799, Option A + confirm-state ordering): the
-// candidate is persisted BEFORE any in-memory promotion AND before any
-// confirm state is touched. On persist failure nothing changes: no
-// promotion, no timer armed, and an EXISTING pending confirm (its
-// timer and rollback target) is left fully intact — the prior ordering
-// cancelled the pending timer and overwrote the rollback target before
-// the write, so a persist failure could strand a pending confirmed
-// commit with no auto-rollback.
-//
-// Nested confirmed commits (a second CommitConfirmed while one is
-// still pending) PRESERVE the existing confirmPrevTree/confirmPrevCfg:
-// the rollback target must stay the last truly CONFIRMED config.
-// Overwriting it with s.active (the unconfirmed commit-1 tree, as the
-// code did before #1799) meant a commit-2 timeout "rolled back" to the
-// equally-unconfirmed commit 1 and stayed there forever, because
-// commit 1's own timer had been cancelled.
-func (s *Store) CommitConfirmed(minutes int) (*config.Config, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return nil, fmt.Errorf("not in configuration mode")
-	}
-
-	compiled, err := s.compileTree(s.candidate)
-	if err != nil {
-		return nil, fmt.Errorf("commit check failed: %w", err)
-	}
-
-	if minutes <= 0 {
-		minutes = 10
-	}
-
-	// #1799 Option A: persist BEFORE promoting and BEFORE touching
-	// any confirm state (see contract above).
-	if err := s.writeActive(s.candidate); err != nil {
-		return nil, fmt.Errorf("commit confirmed failed: persist active config: %w", err)
-	}
-	s.persistDegraded = false // disk now holds the current config
-	// #1922 step-0: a commit confirmed persists the candidate as the
-	// active config (committed=1 on disk via writeActive). The marker is
-	// set here, NOT gated on confirmation — the on-disk DB is now a real
-	// committed config; if the timer fires, the Item 1b first-commit
-	// rollback path (prevCfg==nil) re-writes the never-committed marker.
-	s.everCommitted = true
-	s.persistMarkerCommitted = true
-
-	if s.confirmTimer != nil {
-		// Nested confirmed commit: cancel the pending timer but keep
-		// the original rollback target (last confirmed config).
-		s.confirmTimer.Stop()
-		s.confirmTimer = nil
-	} else {
-		// Save current active state for potential rollback.
-		s.confirmPrevTree = s.active.Clone()
-		s.confirmPrevCfg = s.compiled
-	}
-
-	// Push current active to history
-	s.history.Push(&HistoryEntry{
-		Config:    s.active.Clone(),
-		Timestamp: time.Now(),
-	})
-
-	// Promote candidate to active
-	s.active = s.candidate
-	s.candidate = s.active.Clone()
-	s.compiled = compiled
-	s.dirty = false
-
-	// Log to journal
-	s.journalLog(&JournalEntry{
-		Action:     "commit_confirmed",
-		ConfigHash: journalConfigHash(s.active),
-	})
-
-	s.saveRollbackFiles()
-
-	// Start auto-rollback timer. The closure captures the generation
-	// at arm time; a stale callback from a superseded timer no-ops in
-	// performAutoRollback.
-	s.confirmGen++
-	gen := s.confirmGen
-	s.confirmTimer = time.AfterFunc(time.Duration(minutes)*time.Minute, func() {
-		s.fireConfirmTimer(gen)
-	})
-
-	slog.Info("commit confirmed started", "timeout_minutes", minutes)
-	return compiled, nil
-}
-
-// ConfirmCommit cancels the auto-rollback timer, confirming the config.
-func (s *Store) ConfirmCommit() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.confirmTimer == nil {
-		return fmt.Errorf("no pending confirmed commit")
-	}
-
-	s.confirmTimer.Stop()
-	s.confirmGen++ // invalidate a callback that already fired and is blocked on s.mu
-	s.confirmTimer = nil
-	s.confirmPrevTree = nil
-	s.confirmPrevCfg = nil
-
-	slog.Info("commit confirmed")
-	return nil
-}
-
-// IsConfirmPending returns true if a commit confirmed is awaiting confirmation.
-func (s *Store) IsConfirmPending() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.confirmTimer != nil
-}
-
-// fireConfirmTimer is the commit-confirmed auto-rollback timer's expiry
-// dispatch (#1922 Item 1a). It runs on the timer's own goroutine.
-//
-// It prefers the daemon-owned rollback transaction (rollbackExecutor) so
-// store promotion + dataplane re-apply run atomically under the daemon
-// apply semaphore — and so SERVICE-mode (gRPC/REST/remote-cli) timeouts
-// re-apply the dataplane at all, which the old interactive-only callback
-// never did. The executor is read under s.mu but invoked WITHOUT holding
-// s.mu (it acquires applySem then re-enters the store via PromoteRollback;
-// holding s.mu across that would invert the applySem->s.mu lock order).
-// When no executor is registered (tests / non-daemon embedders) it falls
-// back to the self-contained performAutoRollback (store-state only).
-func (s *Store) fireConfirmTimer(gen uint64) {
-	s.mu.Lock()
-	exec := s.rollbackExecutor
-	s.mu.Unlock()
-	if exec != nil {
-		exec(gen)
-	} else {
-		s.performAutoRollback(gen)
-	}
-}
-
-// PromoteRollback performs ONLY the store-state half of a commit-confirmed
-// timeout rollback, atomically under s.mu, and returns the compiled
-// pre-confirmed config so the caller can re-apply it to the dataplane
-// (#1922 Item 1a, Path Option B). The store is the single owner of the
-// promotion primitive; the daemon's rollback executor holds the apply
-// semaphore around this call + its own re-apply so promotion and re-apply
-// are one critical section relative to a concurrent commit.
-//
-// gen must be the confirmGen captured when the calling timer was armed.
-// A mismatch means the callback was superseded (nested CommitConfirmed
-// re-armed, or ConfirmCommit confirmed) after it fired but before it took
-// s.mu; rolling back would revert a NEWER commit to an older tree (Codex
-// review on PR #1817). On mismatch — or when there is no pending rollback
-// target (confirmPrevTree==nil, e.g. a stale/double timer fire) — this
-// returns (nil, false) and mutates nothing.
-//
-// IMPORTANT: ok=true means "store state was promoted", NOT "prevCfg is
-// non-nil". On a FIRST commit confirmed (fresh store) the rollback target
-// tree is the empty pre-config tree (confirmPrevTree != nil) but the
-// compiled config recorded at arm time is nil (a fresh store has
-// active=&ConfigTree{} but compiled=nil). PromoteRollback then promotes
-// the store back to the empty tree exactly as the prior performAutoRollback
-// did and returns (nil, true). The caller MUST nil-check prevCfg before
-// applying it to the dataplane — exactly as the old code's
-// `if fn != nil && prevCfg != nil` guard did. Re-applying that
-// first-commit-to-bootstrap case to the dataplane (and not persisting an
-// empty *committed* tree) is #1922 Item 1b, DEFERRED to PR-2; this PR
-// leaves that path's behavior unchanged (store reverts, dataplane is not
-// re-applied). The #1817 confirmGen guard and #1799 persist-failure
-// semantics are preserved verbatim.
-func (s *Store) PromoteRollback(gen uint64) (prevCfg *config.Config, ok bool) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if gen != s.confirmGen {
-		return nil, false
-	}
-	if s.confirmPrevTree == nil {
-		return nil, false
-	}
-
-	s.active = s.confirmPrevTree
-	s.compiled = s.confirmPrevCfg
-	if s.candidate != nil {
-		s.candidate = s.active.Clone()
-	}
-	s.dirty = false
-
-	s.confirmTimer = nil
-	s.confirmPrevTree = nil
-	// #1922 Item 1b: first-commit rollback target. On a fresh-store FIRST
-	// commit confirmed, confirmPrevCfg is nil here, so prevCfg below is nil
-	// and ok is true — the store reverts to the empty tree. The caller
-	// (the daemon executor) detects prevCfg==nil and rolls the dataplane
-	// back via enterBootstrapMode rather than applying an empty config.
-	prevCfg = s.confirmPrevCfg
-	s.confirmPrevCfg = nil
-	firstCommitRollback := prevCfg == nil
-
-	// Persist reverted config to disk. Option B (#1799): the rollback
-	// ALWAYS proceeds in memory — reverting the running config is the
-	// safety property — but a persist failure here used to leave the
-	// UNCONFIRMED candidate on disk, so a reboot would load the config
-	// the operator never confirmed. The degraded flag + singleton
-	// retry make the failure visible (/health 503, gauge, journal
-	// ERROR) and heal the disk in the background.
-	//
-	// #1922 Item 1b: on a first-commit rollback the reverted tree is the
-	// empty bootstrap tree. It MUST be persisted with the never-committed
-	// marker (committed=0), NOT as an operator-committed-empty config —
-	// otherwise a subsequent restart would classify committed-empty =>
-	// normal (Item 2 case 5) and take over interfaces on an empty config.
-	// everCommitted is cleared so the in-memory predicate also reads
-	// never-committed without a restart.
-	var perr error
-	if firstCommitRollback {
-		// Record the never-committed marker for BOTH the immediate write
-		// and the degraded-retry loop, so a failed-then-healed write still
-		// persists committed=0 (Codex r1 release-blocker).
-		s.persistMarkerCommitted = false
-		s.everCommitted = false
-		perr = s.writeActiveMarker(s.active, false)
-	} else {
-		s.persistMarkerCommitted = true
-		perr = s.writeActive(s.active)
-	}
-	if perr != nil {
-		s.noteActivePersistFailureLocked("auto_rollback", perr)
-	} else {
-		s.persistDegraded = false
-	}
-
-	// Log to journal
-	s.journalLog(&JournalEntry{
-		Action:     "auto_rollback",
-		ConfigHash: journalConfigHash(s.active),
-	})
-
-	return prevCfg, true
-}
-
-// performAutoRollback is the self-contained fallback rollback used when no
-// daemon rollback executor is registered (tests / non-daemon embedders).
-// It promotes store state via PromoteRollback (the single promotion
-// primitive); it does NOT re-apply the dataplane — a non-daemon embedder
-// has no dataplane to re-apply, and the daemon path goes through
-// SetRollbackExecutor + executeConfirmedRollback instead (#1922 Item 1a).
-// gen must be the confirmGen captured when the calling timer was armed.
-func (s *Store) performAutoRollback(gen uint64) {
-	if _, ok := s.PromoteRollback(gen); !ok {
-		return
-	}
-	slog.Warn("commit confirmed timed out, configuration rolled back")
-}
-
-// Rollback reverts the candidate to a previous configuration.
-// n=0 reverts to active; n>0 reverts to the nth previous commit.
-func (s *Store) Rollback(n int) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.candidate == nil {
-		return fmt.Errorf("not in configuration mode")
-	}
-
-	if n == 0 {
-		s.candidate = s.active.Clone()
-		s.dirty = false
-		return nil
-	}
-
-	entry, err := s.history.Get(n - 1)
-	if err != nil {
-		return err
-	}
-	s.candidate = entry.Config.Clone()
-	s.dirty = true
-	return nil
-}
-
-// ShowCandidate returns the candidate configuration as hierarchical text.
-func (s *Store) ShowCandidate() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.Format()
-	}
-	return ""
-}
-
-// ShowCandidatePath returns the candidate configuration subtree at the given path.
-func (s *Store) ShowCandidatePath(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatPath(path)
-	}
-	return ""
-}
-
-// ShowActive returns the active configuration as hierarchical text.
-func (s *Store) ShowActive() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.Format()
-}
-
-// ShowActivePath returns the active configuration subtree at the given path.
-func (s *Store) ShowActivePath(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatPath(path)
-}
-
-// ShowCandidateSet returns the candidate configuration as flat set commands.
-func (s *Store) ShowCandidateSet() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatSet()
-	}
-	return ""
-}
-
-// ActiveConfig returns the compiled active configuration.
-func (s *Store) ActiveConfig() *config.Config {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.compiled
-}
-
-// CompileCandidate strictly compiles the current candidate WITHOUT mutating
-// any store state (no promote, no persist, no confirm-timer). It is the
-// read-only pre-commit hook the daemon's #1956 device-map commit pre-flight
-// uses to resolve the proposed map against live hardware BEFORE the store
-// promotes it — so a map that would strand management on next boot is
-// rejected while the operator is still connected, not at the next reboot.
-// Returns the same compiled config Commit() would, or the commit-check error.
-func (s *Store) CompileCandidate() (*config.Config, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate == nil {
-		return nil, fmt.Errorf("not in configuration mode")
-	}
-	return s.compileTree(s.candidate)
-}
-
-// EverCommitted reports the #1922 step-0 marker: true once a config has
-// been successfully committed or synced to this store, or loaded from a
-// committed/legacy DB (migration rule C3 defaults a marker-less DB to
-// committed). It is false on a fresh store and after the Item 1b
-// first-commit rollback persists the never-committed marker. The daemon's
-// five-case boot predicate reads it to disambiguate operator-committed-empty
-// (normal) from never-committed (bootstrap) when the active config is empty.
-func (s *Store) EverCommitted() bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.everCommitted
-}
-
-// ActiveTree returns a deep copy of the active configuration tree.
-func (s *Store) ActiveTree() *config.ConfigTree {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.active == nil {
-		return nil
-	}
-	return s.active.Clone()
-}
-
-// ExportJSON exports the active config as JSON (for debugging).
-func (s *Store) ExportJSON() ([]byte, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return json.MarshalIndent(s.compiled, "", "  ")
-}
-
-// ListHistory returns all history entries, most recent first (goroutine-safe).
-func (s *Store) ListHistory() []*HistoryEntry {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.history.List()
-}
-
-// ListCommitHistory returns recent commit journal entries (most recent
-// last). The journal read is bounded by limit (#1896): the tail scan
-// stops after limit entries, so cost is O(limit), not O(lifetime
-// journal). Semantics preserved from v1: the last `limit` entries of
-// ANY action are read first, THEN filtered to commit actions — fewer
-// than `limit` commits can come back when other actions interleave.
-// limit <= 0 reads everything (persist_failure_test relies on it).
-// No Store.mu needed: the journal serializes Log/Tail internally.
-func (s *Store) ListCommitHistory(limit int) ([]*JournalEntry, error) {
-	entries, err := s.journal.Tail(limit)
-	if err != nil {
-		return nil, err
-	}
-	// Filter to commit/rollback actions only
-	var commits []*JournalEntry
-	for _, e := range entries {
-		switch e.Action {
-		case "commit", "commit_confirmed", "auto_rollback":
-			commits = append(commits, e)
-		}
-	}
-	return commits, nil
-}
-
-// CommitDiffSummary returns a human-readable summary of changes between
-// the active and candidate configs. Must be called while in config mode.
-func (s *Store) CommitDiffSummary() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	if s.candidate == nil {
-		return ""
-	}
-
-	activeSet := s.active.FormatSet()
-	candidateSet := s.candidate.FormatSet()
-
-	activeLines := splitLines(activeSet)
-	candidateLines := splitLines(candidateSet)
-
-	activeMap := make(map[string]bool, len(activeLines))
-	for _, line := range activeLines {
-		activeMap[line] = true
-	}
-	candidateMap := make(map[string]bool, len(candidateLines))
-	for _, line := range candidateLines {
-		candidateMap[line] = true
-	}
-
-	var added, removed int
-	for _, line := range activeLines {
-		if !candidateMap[line] {
-			removed++
-		}
-	}
-	for _, line := range candidateLines {
-		if !activeMap[line] {
-			added++
-		}
-	}
-
-	total := added + removed
-	if total == 0 {
-		return ""
-	}
-
-	return fmt.Sprintf("%d statement(s) changed (%d added, %d removed)", total, added, removed)
-}
-
-// rollbackPath returns the file path for rollback slot n (1-based).
-func (s *Store) rollbackPath(n int) string {
-	return filepath.Join(filepath.Dir(s.filePath), fmt.Sprintf("%s.%d", filepath.Base(s.filePath), n))
-}
-
-// saveRollbackFiles writes rollback history entries to numbered files.
-// Must be called under write lock.
-//
-// Durability split (#1894, adjudicated in the plan round): these files
-// are the CANONICAL rollback history (loadRollbackHistory reads them at
-// boot; the DB rollback slots have no production callers). Slot 1 — the
-// immediate `rollback 1` target — is written durably; slots 2..N use
-// the atomic writer (never missing, never torn, so loadRollbackHistory's
-// break-on-first-missing stays sound; they may lag behind after a power
-// cut). One trailing SyncDir then makes the whole shuffle AND the
-// stale-slot unlinks durable for the cost of a single dir fsync,
-// instead of ~50 file+dir fsync pairs under the store mutex.
-func (s *Store) saveRollbackFiles() {
-	if s.filePath == "" {
-		return
-	}
-
-	entries := s.history.List() // most-recent-first
-	for i, entry := range entries {
-		path := s.rollbackPath(i + 1)
-		data := entry.Config.Format()
-		var err error
-		if i == 0 {
-			err = fsatomic.WriteFileDurable(path, []byte(data), 0644)
-		} else {
-			err = fsatomic.WriteFileAtomic(path, []byte(data), 0644)
-		}
-		if err != nil {
-			slog.Warn("failed to write rollback file", "path", path, "err", err)
-		}
-	}
-	s.cleanupRollbackFiles(len(entries) + 1)
-	if len(entries) > 0 {
-		if err := fsatomic.SyncDir(filepath.Dir(s.filePath)); err != nil {
-			slog.Warn("failed to sync rollback directory", "err", err)
-		}
-	}
-}
-
-// cleanupRollbackFiles removes stale rollback files starting at startN.
-func (s *Store) cleanupRollbackFiles(startN int) {
-	for i := startN; i <= s.history.MaxSize()+1; i++ {
-		path := s.rollbackPath(i)
-		if err := os.Remove(path); err != nil {
-			break // stop on first not-found (contiguous sequence)
-		}
-	}
-}
-
-// loadRollbackHistory reads numbered rollback files and populates the history.
-// Must be called under write lock.
-func (s *Store) loadRollbackHistory() {
-	if s.filePath == "" {
-		return
-	}
-
-	var entries []*HistoryEntry
-	for i := 1; i <= s.history.MaxSize(); i++ {
-		path := s.rollbackPath(i)
-		data, err := os.ReadFile(path)
-		if err != nil {
-			break // stop on first not-found
-		}
-		parser := config.NewParser(string(data))
-		tree, errs := parser.Parse()
-		if len(errs) > 0 {
-			slog.Warn("skipping corrupt rollback file", "path", path, "err", errs[0])
-			continue
-		}
-		// Use file modification time as timestamp
-		info, _ := os.Stat(path)
-		ts := time.Now()
-		if info != nil {
-			ts = info.ModTime()
-		}
-		entries = append(entries, &HistoryEntry{
-			Config:    tree,
-			Timestamp: ts,
-		})
-	}
-
-	// Push oldest-first so History ordering is correct
-	for i := len(entries) - 1; i >= 0; i-- {
-		s.history.Push(entries[i])
-	}
-
-	if len(entries) > 0 {
-		slog.Info("loaded rollback history", "entries", len(entries))
-	}
-}
-
-// ShowActiveSet returns the active configuration as flat set commands.
-func (s *Store) ShowActiveSet() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatSet()
-}
-
-// ShowActivePathSet returns an active config subtree as flat set commands.
-func (s *Store) ShowActivePathSet(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatPathSet(path)
-}
-
-// ShowCandidatePathSet returns a candidate config subtree as flat set commands.
-func (s *Store) ShowCandidatePathSet(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatPathSet(path)
-	}
-	return ""
-}
-
-// ShowActiveJSON returns the active configuration as JSON.
-func (s *Store) ShowActiveJSON() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatJSON()
-}
-
-// ShowActivePathJSON returns an active config subtree as JSON.
-func (s *Store) ShowActivePathJSON(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatPathJSON(path)
-}
-
-// ShowCandidateJSON returns the candidate configuration as JSON.
-func (s *Store) ShowCandidateJSON() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatJSON()
-	}
-	return "{}\n"
-}
-
-// ShowCandidatePathJSON returns a candidate config subtree as JSON.
-func (s *Store) ShowCandidatePathJSON(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatPathJSON(path)
-	}
-	return ""
-}
-
-// ShowActiveXML returns the active configuration as XML.
-func (s *Store) ShowActiveXML() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatXML()
-}
-
-// ShowActivePathXML returns an active config subtree as XML.
-func (s *Store) ShowActivePathXML(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatPathXML(path)
-}
-
-// ShowCandidateXML returns the candidate configuration as XML.
-func (s *Store) ShowCandidateXML() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatXML()
-	}
-	return "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<configuration>\n</configuration>\n"
-}
-
-// ShowCandidatePathXML returns a candidate config subtree as XML.
-func (s *Store) ShowCandidatePathXML(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatPathXML(path)
-	}
-	return ""
-}
-
-// ShowCandidateInheritance returns the candidate with groups expanded and
-// annotated with "## inherited from" comments.
-func (s *Store) ShowCandidateInheritance() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatInheritance()
-	}
-	return ""
-}
-
-// ShowCandidatePathInheritance returns a subtree with inheritance annotations.
-func (s *Store) ShowCandidatePathInheritance(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	if s.candidate != nil {
-		return s.candidate.FormatPathInheritance(path)
-	}
-	return ""
-}
-
-// ShowActiveInheritance returns the active config with inheritance annotations.
-func (s *Store) ShowActiveInheritance() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatInheritance()
-}
-
-// ShowActivePathInheritance returns an active config subtree with inheritance annotations.
-func (s *Store) ShowActivePathInheritance(path []string) string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.active.FormatPathInheritance(path)
-}
-
-// ShowRollback returns the content of rollback slot n (1-based) as hierarchical text.
-func (s *Store) ShowRollback(n int) (string, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	entry, err := s.history.Get(n - 1)
-	if err != nil {
-		return "", err
-	}
-	return entry.Config.Format(), nil
-}
-
-// ShowRollbackSet returns the content of rollback slot n (1-based) as flat set commands.
-func (s *Store) ShowRollbackSet(n int) (string, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	entry, err := s.history.Get(n - 1)
-	if err != nil {
-		return "", err
-	}
-	return entry.Config.FormatSet(), nil
-}
-
-// ShowCompareRollback returns a diff between rollback slot n and the candidate.
-func (s *Store) ShowCompareRollback(n int) (string, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	if s.candidate == nil {
-		return "", fmt.Errorf("not in configuration mode")
-	}
-
-	entry, err := s.history.Get(n - 1)
-	if err != nil {
-		return "", err
-	}
-
-	diff := config.FormatCompare(entry.Config, s.candidate)
-	if diff == "" {
-		return "[no changes]\n", nil
-	}
-	return diff, nil
-}
-
-// ShowCompare returns a hierarchical diff between the active and candidate
-// configurations in Junos [edit] context format.
-func (s *Store) ShowCompare() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	if s.candidate == nil {
-		return ""
-	}
-
-	diff := config.FormatCompare(s.active, s.candidate)
-	if diff == "" {
-		return "[no changes]\n"
-	}
-	return diff
-}
-
-// splitLines splits a string into non-empty lines.
-func splitLines(s string) []string {
-	var lines []string
-	for _, line := range strings.Split(s, "\n") {
-		line = strings.TrimSpace(line)
-		if line != "" {
-			lines = append(lines, line)
-		}
-	}
-	return lines
-}
-
-// SetArchiveConfig configures automatic archival on commit.
-func (s *Store) SetArchiveConfig(dir string, max int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.archiveDir = dir
-	s.archiveMax = max
-}
-
-// ArchiveConfig saves a timestamped copy of the active config.
-func (s *Store) ArchiveConfig(archiveDir string, maxArchives int) error {
-	s.mu.RLock()
-	data := s.active.Format()
-	s.mu.RUnlock()
-
-	if err := os.MkdirAll(archiveDir, 0755); err != nil {
-		return fmt.Errorf("create archive dir: %w", err)
-	}
-
-	filename := fmt.Sprintf("config-%s.conf", time.Now().Format("20060102-150405"))
-	path := filepath.Join(archiveDir, filename)
-	// AtomicGeneratedConfig (#1894): archives are best-effort history
-	// copies — atomic so a crash never leaves a torn archive, but not
-	// worth an fsync.
-	if err := fsatomic.WriteFileAtomic(path, []byte(data), 0644); err != nil {
-		return fmt.Errorf("write archive: %w", err)
-	}
-
-	slog.Info("config archived", "path", path)
-
-	// Rotate old archives
-	if maxArchives > 0 {
-		rotateArchives(archiveDir, maxArchives)
-	}
-	return nil
-}
-
-// rotateArchives keeps only the most recent maxArchives files.
-func rotateArchives(dir string, maxArchives int) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return
-	}
-
-	var archives []string
-	for _, e := range entries {
-		if !e.IsDir() && strings.HasPrefix(e.Name(), "config-") && strings.HasSuffix(e.Name(), ".conf") {
-			archives = append(archives, e.Name())
-		}
-	}
-
-	if len(archives) <= maxArchives {
-		return
-	}
-
-	// Sort alphabetically (timestamps sort naturally)
-	sort.Strings(archives)
-
-	// Remove oldest
-	for i := 0; i < len(archives)-maxArchives; i++ {
-		path := filepath.Join(dir, archives[i])
-		if err := os.Remove(path); err != nil {
-			slog.Warn("failed to remove old archive", "path", path, "err", err)
-		}
-	}
-}
-
-// rescuePath returns the path for the rescue configuration file.
-func (s *Store) rescuePath() string {
-	return filepath.Join(filepath.Dir(s.filePath), "rescue.conf")
-}
-
-// SaveRescueConfig saves the active config as rescue configuration.
-func (s *Store) SaveRescueConfig() error {
-	s.mu.RLock()
-	data := s.active.Format()
-	s.mu.RUnlock()
-
-	path := s.rescuePath()
-	// DurableState (#1894): the rescue config is the operator's
-	// explicitly-requested safety net — it must survive power loss.
-	if err := fsatomic.WriteFileDurable(path, []byte(data), 0644); err != nil {
-		return fmt.Errorf("save rescue config: %w", err)
-	}
-	slog.Info("rescue configuration saved", "path", path)
-	return nil
-}
-
-// DeleteRescueConfig removes the rescue configuration.
-func (s *Store) DeleteRescueConfig() error {
-	path := s.rescuePath()
-	if err := os.Remove(path); err != nil {
-		if os.IsNotExist(err) {
-			return fmt.Errorf("no rescue configuration exists")
-		}
-		return fmt.Errorf("delete rescue config: %w", err)
-	}
-	slog.Info("rescue configuration deleted", "path", path)
-	return nil
-}
-
-// LoadRescueConfig returns the rescue configuration text, or "" if none.
-func (s *Store) LoadRescueConfig() (string, error) {
-	path := s.rescuePath()
-	data, err := os.ReadFile(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return "", nil
-		}
-		return "", fmt.Errorf("read rescue config: %w", err)
-	}
-	return string(data), nil
 }
